trigger notion import on load
